Move catalog verification into the sys pacakge for better reuse
