+++ pkg/pe/marshaller.go
@@ -24,6 +24,7 @@ package pe
 import (
 	"fmt"
 	kcapver "github.com/rabbitstack/fibratus/pkg/kcap/version"
+	"github.com/rabbitstack/fibratus/pkg/sys"
 	"github.com/rabbitstack/fibratus/pkg/util/bytes"
 	"github.com/rabbitstack/fibratus/pkg/util/convert"
 	"math"
@@ -246,7 +247,7 @@ func (pe *PE) Unmarshal(b []byte, ver kcapver.Version) error {
 
 		certSize := bytes.ReadUint32(b[22+offset:])
 		if certSize > 0 {
-			pe.Cert = &Cert{}
+			pe.Cert = &sys.Cert{}
 			err := pe.Cert.Unmarshal(b, offset, certSize)
 			if err != nil {
 				return err
@@ -277,73 +278,6 @@ func (pe *PE) Unmarshal(b []byte, ver kcapver.Version) error {
 			off += 2 + uint32(l)
 		}
 	}
-
-	return nil
-}
-
-// Marshal writes certificate info into a raw buffer.
-func (c *Cert) Marshal() []byte {
-	b := make([]byte, 0)
-
-	before := make([]byte, 0)
-	before = c.NotBefore.AppendFormat(before, time.RFC3339Nano)
-	b = append(b, bytes.WriteUint16(uint16(len(before)))...)
-	b = append(b, before...)
-
-	after := make([]byte, 0)
-	after = c.NotAfter.AppendFormat(after, time.RFC3339Nano)
-	b = append(b, bytes.WriteUint16(uint16(len(after)))...)
-	b = append(b, after...)
-
-	b = append(b, bytes.WriteUint16(uint16(len(c.SerialNumber)))...)
-	b = append(b, c.SerialNumber...)
-	b = append(b, bytes.WriteUint16(uint16(len(c.Subject)))...)
-	b = append(b, c.Subject...)
-	b = append(b, bytes.WriteUint16(uint16(len(c.Issuer)))...)
-	b = append(b, c.Issuer...)
-
-	return b
-}
-
-// Unmarshal decodes cert info from the raw buffer. This method
-// assumes the certificate structure size was already read.
-func (c *Cert) Unmarshal(b []byte, offset, certSize uint32) error {
-	if certSize > uint32(len(b)) {
-		return fmt.Errorf("invalid PE cert size. Got %d but max buffer size is %d", certSize, len(b))
-	}
-
-	// read not before
-	l := bytes.ReadUint16(b[26+offset:])
-	buf := b[28+offset:]
-	offset += uint32(l)
-	if len(buf) > 0 {
-		c.NotBefore, _ = time.Parse(time.RFC3339Nano, string((*[1<<30 - 1]byte)(unsafe.Pointer(&buf[0]))[:l:l]))
-	}
-
-	// read not after
-	l = bytes.ReadUint16(b[28+offset:])
-	buf = b[30+offset:]
-	offset += uint32(l)
-	if len(buf) > 0 {
-		c.NotAfter, _ = time.Parse(time.RFC3339Nano, string((*[1<<30 - 1]byte)(unsafe.Pointer(&buf[0]))[:l:l]))
-	}
-
-	// read serial
-	l = bytes.ReadUint16(b[30+offset:])
-	buf = b[32+offset:]
-	offset += uint32(l)
-	c.SerialNumber = string((*[1<<30 - 1]byte)(unsafe.Pointer(&buf[0]))[:l:l])
-
-	// read subject
-	l = bytes.ReadUint16(b[32+offset:])
-	buf = b[34+offset:]
-	offset += uint32(l)
-	c.Subject = string((*[1<<30 - 1]byte)(unsafe.Pointer(&buf[0]))[:l:l])
-
-	// read issuer
-	l = bytes.ReadUint16(b[34+offset:])
-	buf = b[36+offset:]
-	c.Issuer = string((*[1<<30 - 1]byte)(unsafe.Pointer(&buf[0]))[:l:l])
 
 	return nil
 }
