+++ pkg/sys/trust.go
@@ -19,10 +19,20 @@
 package sys
 
 import (
+	"encoding/hex"
+	"fmt"
+	"github.com/rabbitstack/fibratus/pkg/util/bytes"
 	"github.com/rabbitstack/fibratus/pkg/util/format"
+	"github.com/rabbitstack/fibratus/pkg/util/multierror"
 	log "github.com/sirupsen/logrus"
+	"go.mozilla.org/pkcs7"
 	"golang.org/x/sys/windows"
+	"io"
+	"os"
+	"reflect"
+	"runtime"
 	"sync"
+	"time"
 	"unsafe"
 )
 
@@ -212,4 +222,256 @@ func IsWintrustFound() bool {
 		}
 	})
 	return isWintrustDLLFound
+}
+
+// Cat represents the catalog that acts as a digital signature
+// for an arbitrary collection of files. A catalog file contains
+// a collection of cryptographic hashes, or thumbprints. Each thumbprint
+// corresponds to a file that is included in the collection.
+type Cat struct {
+	admin       windows.Handle
+	catalog     windows.Handle
+	file        *os.File
+	catalogInfo CatalogInfo
+
+	hash []byte
+	size uint32
+}
+
+const hashSize uint32 = 100
+
+// NewCatalog creates an instance of the catalog with default hash size.
+func NewCatalog() Cat {
+	return Cat{size: hashSize, hash: make([]byte, hashSize)}
+}
+
+// Open opens the catalog and acquires the hash for the given file. If the
+// file is catalog-signed, a valid catalog handle is stored internally.
+func (c *Cat) Open(filename string) error {
+	runtime.LockOSThread()
+	defer runtime.UnlockOSThread()
+	// acquire handle to a catalog administrator context
+	err := CryptCatalogAdminAcquireContext(&c.admin, nil, nil, 0, 0)
+	if err != nil {
+		return err
+	}
+	// calculate file hash
+	c.file, err = os.Open(filename)
+	if err != nil {
+		return err
+	}
+	err = CryptCatalogAdminCalcHashFromFileHandle(
+		c.admin,
+		c.file.Fd(),
+		&c.size,
+		uintptr(unsafe.Pointer(&c.hash[0])), 0,
+	)
+	if err != nil {
+		return err
+	}
+	// enumerate catalogs that contain the calculated hash.
+	// If no catalogs are found, we can deduce the file is
+	// not catalog signed
+	c.catalog = CryptCatalogAdminEnumCatalogFromHash(
+		c.admin,
+		uintptr(unsafe.Pointer(&c.hash[0])),
+		c.size, 0, nil,
+	)
+	c.catalogInfo.Size = uint32(unsafe.Sizeof(c.catalogInfo))
+	err = CryptCatalogInfoFromContext(c.catalog, &c.catalogInfo, 0)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+// IsCatalogSigned determines if the file is catalog-signed.
+func (c *Cat) IsCatalogSigned() bool {
+	return c.catalog != 0 && c.catalogInfo.CatalogFile() != ""
+}
+
+// Verify verifies the signature of the given file against the catalog.
+func (c *Cat) Verify(filename string) bool {
+	runtime.LockOSThread()
+	defer runtime.UnlockOSThread()
+	trust := NewWintrustData(WtdChoiceCatalog)
+	defer trust.Close()
+	if c.file == nil {
+		panic("catalog is not opened")
+	}
+	return trust.VerifyCatalog(c.file.Fd(), filename, c.admin, c.catalogInfo, c.hash, c.size)
+}
+
+// ParseCertificate parses the catalog certificate.
+func (c *Cat) ParseCertificate() (*Cert, error) {
+	f, err := os.Open(c.catalogInfo.CatalogFile())
+	if err != nil {
+		return nil, err
+	}
+	defer f.Close()
+	crt, err := io.ReadAll(f)
+	if err != nil {
+		return nil, err
+	}
+	pkcs, err := pkcs7.Parse(crt)
+	if err != nil {
+		return nil, err
+	}
+
+	certInfo := &Cert{}
+	serialNumber := pkcs.Signers[0].IssuerAndSerialNumber.SerialNumber
+	for _, cert := range pkcs.Certificates {
+		if !reflect.DeepEqual(cert.SerialNumber, serialNumber) {
+			continue
+		}
+
+		certInfo.SerialNumber = hex.EncodeToString(cert.SerialNumber.Bytes())
+
+		certInfo.NotAfter = cert.NotAfter
+		certInfo.NotBefore = cert.NotBefore
+
+		// issuer information
+		if len(cert.Issuer.Country) > 0 {
+			certInfo.Issuer = cert.Issuer.Country[0]
+		}
+
+		if len(cert.Issuer.Province) > 0 {
+			certInfo.Issuer += ", " + cert.Issuer.Province[0]
+		}
+
+		if len(cert.Issuer.Locality) > 0 {
+			certInfo.Issuer += ", " + cert.Issuer.Locality[0]
+		}
+
+		certInfo.Issuer += ", " + cert.Issuer.CommonName
+
+		// subject information
+		if len(cert.Subject.Country) > 0 {
+			certInfo.Subject = cert.Subject.Country[0]
+		}
+
+		if len(cert.Subject.Province) > 0 {
+			certInfo.Subject += ", " + cert.Subject.Province[0]
+		}
+
+		if len(cert.Subject.Locality) > 0 {
+			certInfo.Subject += ", " + cert.Subject.Locality[0]
+		}
+
+		if len(cert.Subject.Organization) > 0 {
+			certInfo.Subject += ", " + cert.Subject.Organization[0]
+		}
+
+		certInfo.Subject += ", " + cert.Subject.CommonName
+
+		break
+	}
+	return certInfo, nil
+}
+
+func (c *Cat) Close() error {
+	runtime.LockOSThread()
+	defer runtime.UnlockOSThread()
+	if c.admin != 0 {
+		defer CryptCatalogAdminReleaseContext(c.admin, 0)
+	}
+	var err error
+	if c.admin != 0 && c.catalog != 0 {
+		err = CryptCatalogAdminReleaseCatalogContext(c.admin, c.catalog, 0)
+	}
+	if c.file != nil {
+		return multierror.Wrap(c.file.Close(), err)
+	}
+	return err
+}
+
+// Cert represents certificate information embedded in the PE or catalog.
+type Cert struct {
+	// NotBefore specifies the certificate won't be valid before this timestamp.
+	NotBefore time.Time `json:"not_before"`
+
+	// NotAfter specifies the certificate won't be valid after this timestamp.
+	NotAfter time.Time `json:"not_after"`
+
+	// Issuer represents the certificate authority (CA) that charges customers to issue
+	// certificates for them.
+	Issuer string `json:"issuer"`
+
+	// Subject indicates the subject of the certificate is the entity its public key is associated
+	// with (i.e. the "owner" of the certificate).
+	Subject string `json:"subject"`
+
+	// SerialNumber represents the serial number MUST be a positive integer assigned
+	// by the CA to each certificate. It MUST be unique for each certificate issued by
+	// a given CA (i.e., the issuer name and serial number identify a unique certificate).
+	// CAs MUST force the serialNumber to be a non-negative integer.
+	// For convenience, we convert the big int to string.
+	SerialNumber string `json:"serial_number"`
+}
+
+// Marshal writes certificate info into a raw buffer.
+func (c *Cert) Marshal() []byte {
+	b := make([]byte, 0)
+
+	before := make([]byte, 0)
+	before = c.NotBefore.AppendFormat(before, time.RFC3339Nano)
+	b = append(b, bytes.WriteUint16(uint16(len(before)))...)
+	b = append(b, before...)
+
+	after := make([]byte, 0)
+	after = c.NotAfter.AppendFormat(after, time.RFC3339Nano)
+	b = append(b, bytes.WriteUint16(uint16(len(after)))...)
+	b = append(b, after...)
+
+	b = append(b, bytes.WriteUint16(uint16(len(c.SerialNumber)))...)
+	b = append(b, c.SerialNumber...)
+	b = append(b, bytes.WriteUint16(uint16(len(c.Subject)))...)
+	b = append(b, c.Subject...)
+	b = append(b, bytes.WriteUint16(uint16(len(c.Issuer)))...)
+	b = append(b, c.Issuer...)
+
+	return b
+}
+
+// Unmarshal decodes cert info from the raw buffer. This method
+// assumes the certificate structure size was already read.
+func (c *Cert) Unmarshal(b []byte, offset, certSize uint32) error {
+	if certSize > uint32(len(b)) {
+		return fmt.Errorf("invalid PE cert size. Got %d but max buffer size is %d", certSize, len(b))
+	}
+
+	// read not before
+	l := bytes.ReadUint16(b[26+offset:])
+	buf := b[28+offset:]
+	offset += uint32(l)
+	if len(buf) > 0 {
+		c.NotBefore, _ = time.Parse(time.RFC3339Nano, string((*[1<<30 - 1]byte)(unsafe.Pointer(&buf[0]))[:l:l]))
+	}
+
+	// read not after
+	l = bytes.ReadUint16(b[28+offset:])
+	buf = b[30+offset:]
+	offset += uint32(l)
+	if len(buf) > 0 {
+		c.NotAfter, _ = time.Parse(time.RFC3339Nano, string((*[1<<30 - 1]byte)(unsafe.Pointer(&buf[0]))[:l:l]))
+	}
+
+	// read serial
+	l = bytes.ReadUint16(b[30+offset:])
+	buf = b[32+offset:]
+	offset += uint32(l)
+	c.SerialNumber = string((*[1<<30 - 1]byte)(unsafe.Pointer(&buf[0]))[:l:l])
+
+	// read subject
+	l = bytes.ReadUint16(b[32+offset:])
+	buf = b[34+offset:]
+	offset += uint32(l)
+	c.Subject = string((*[1<<30 - 1]byte)(unsafe.Pointer(&buf[0]))[:l:l])
+
+	// read issuer
+	l = bytes.ReadUint16(b[34+offset:])
+	buf = b[36+offset:]
+	c.Issuer = string((*[1<<30 - 1]byte)(unsafe.Pointer(&buf[0]))[:l:l])
+
+	return nil
 }
